pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.76.0",
    "google-generativeai>=0.8.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "rebound>=4.5.0",
]
